python--kafka-produce-consume/requirements.txt
kafka-python
pyyaml
